requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
colorama==0.4.6
distlib==0.3.9
execnet==2.1.1
faker==30.8.1
filelock==3.16.1
greenlet==3.0.3
identify==2.6.2
idna==3.10
iniconfig==2.0.0
nodeenv==1.9.1
packaging==24.1
pip==24.3.1
platformdirs==4.3.6
playwright==1.47.0
pluggy==1.5.0
pre-commit==4.0.1
pyee==12.0.0
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyyaml==6.0.2
requests==2.32.3
ruff==0.7.4
six==1.16.0
text-unidecode==1.3
typing-extensions==4.12.2
urllib3==2.2.3
uv==0.5.2
virtualenv==20.27.1
